code cleanup and institute proper hierarchy
