+++ benchmarks-jmh/src/main/java/io/github/jbellis/jvector/bench/RandomVectorsBenchmark.java
@@ -17,19 +17,14 @@ package io.github.jbellis.jvector.bench;
 
 import io.github.jbellis.jvector.example.SiftSmall;
 import io.github.jbellis.jvector.graph.*;
-import io.github.jbellis.jvector.graph.similarity.BuildScoreProvider;
 import io.github.jbellis.jvector.util.Bits;
 import io.github.jbellis.jvector.vector.VectorSimilarityFunction;
-import io.github.jbellis.jvector.vector.VectorizationProvider;
-import io.github.jbellis.jvector.vector.types.VectorFloat;
-import io.github.jbellis.jvector.vector.types.VectorTypeSupport;
 import org.openjdk.jmh.annotations.*;
 import org.openjdk.jmh.infra.Blackhole;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.concurrent.TimeUnit;
 
 @BenchmarkMode(Mode.AverageTime)
@@ -39,15 +34,9 @@ import java.util.concurrent.TimeUnit;
 @Warmup(iterations = 2)
 @Measurement(iterations = 5)
 @Threads(1)
-public class RandomVectorsBenchmark {
+public class RandomVectorsBenchmark extends AbstractVectorsBenchmark {
     private static final Logger log = LoggerFactory.getLogger(RandomVectorsBenchmark.class);
-    private static final VectorTypeSupport VECTOR_TYPE_SUPPORT = VectorizationProvider.getInstance().getVectorTypeSupport();
-    private RandomAccessVectorValues ravv;
-    private ArrayList<VectorFloat<?>> baseVectors;
-    private ArrayList<VectorFloat<?>> queryVectors;
-    private GraphIndexBuilder graphIndexBuilder;
-    private GraphIndex graphIndex;
-    int originalDimension;
+
     @Param({"1000", "10000", "100000", "1000000"})
     int numBaseVectors;
     @Param({"10"})
@@ -55,42 +44,7 @@ public class RandomVectorsBenchmark {
 
     @Setup
     public void setup() throws IOException {
-        originalDimension = 128; // Example dimension, can be adjusted
-
-        baseVectors = new ArrayList<>(numBaseVectors);
-        queryVectors = new ArrayList<>(numQueryVectors);
-
-        for (int i = 0; i < numBaseVectors; i++) {
-            VectorFloat<?> vector = createRandomVector(originalDimension);
-            baseVectors.add(vector);
-        }
-
-        for (int i = 0; i < numQueryVectors; i++) {
-            VectorFloat<?> vector = createRandomVector(originalDimension);
-            queryVectors.add(vector);
-        }
-
-        // wrap the raw vectors in a RandomAccessVectorValues
-        ravv = new ListRandomAccessVectorValues(baseVectors, originalDimension);
-
-        // score provider using the raw, in-memory vectors
-        BuildScoreProvider bsp = BuildScoreProvider.randomAccessScoreProvider(ravv, VectorSimilarityFunction.EUCLIDEAN);
-
-        graphIndexBuilder = new GraphIndexBuilder(bsp,
-                ravv.dimension(),
-                16, // graph degree
-                100, // construction search depth
-                1.2f, // allow degree overflow during construction by this factor
-                1.2f); // relax neighbor diversity requirement by this factor
-        graphIndex = graphIndexBuilder.build(ravv);
-    }
-
-    private VectorFloat<?> createRandomVector(int dimension) {
-        VectorFloat<?> vector = VECTOR_TYPE_SUPPORT.createFloatVector(dimension);
-        for (int i = 0; i < dimension; i++) {
-            vector.set(i, (float) Math.random());
-        }
-        return vector;
+        commonSetupRandom(numBaseVectors, numQueryVectors);
     }
 
     @TearDown
