+++ benchmarks-jmh/src/main/java/io/github/jbellis/jvector/bench/RandomVectorsTabularBenchmark.java
@@ -1,22 +1,31 @@
+/*
+ * Copyright DataStax, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.github.jbellis.jvector.bench;
 
-import io.github.jbellis.jvector.bench.output.TableRepresentation;
 import io.github.jbellis.jvector.bench.output.TextTable;
 import io.github.jbellis.jvector.example.SiftSmall;
 import io.github.jbellis.jvector.graph.*;
-import io.github.jbellis.jvector.graph.similarity.BuildScoreProvider;
 import io.github.jbellis.jvector.util.Bits;
 import io.github.jbellis.jvector.vector.VectorSimilarityFunction;
-import io.github.jbellis.jvector.vector.VectorizationProvider;
-import io.github.jbellis.jvector.vector.types.VectorFloat;
-import io.github.jbellis.jvector.vector.types.VectorTypeSupport;
 import org.openjdk.jmh.annotations.*;
 import org.openjdk.jmh.infra.Blackhole;
 
 import java.io.IOException;
-import java.util.*;
 import java.util.concurrent.*;
-import java.util.concurrent.atomic.AtomicLong;
 
 @BenchmarkMode(Mode.Throughput)
 @OutputTimeUnit(TimeUnit.SECONDS)
@@ -25,27 +34,7 @@ import java.util.concurrent.atomic.AtomicLong;
 @Warmup(iterations = 2)
 @Measurement(iterations = 5)
 @Threads(1)
-public class RandomVectorsTabularBenchmark {
-    private static final VectorTypeSupport VECTOR_TYPE_SUPPORT = VectorizationProvider.getInstance().getVectorTypeSupport();
-    private RandomAccessVectorValues ravv;
-
-    private ArrayList<VectorFloat<?>> baseVectors;
-    private ArrayList<VectorFloat<?>> queryVectors;
-    private GraphIndexBuilder graphIndexBuilder;
-    private GraphIndex graphIndex;
-    private int originalDimension;
-    private long totalTransactions;
-
-    private final AtomicLong transactionCount = new AtomicLong(0);
-    private final AtomicLong totalLatency = new AtomicLong(0);
-    private final Queue<Long> latencySamples = new ConcurrentLinkedQueue<>(); // Store latencies for P99.9
-    private final Queue<Integer> visitedSamples = new ConcurrentLinkedQueue<>();
-    private final Queue<Long> recallSamples = new ConcurrentLinkedQueue<>();
-    private ScheduledExecutorService scheduler;
-    private long startTime;
-
-    private final TableRepresentation tableRepresentation = new TextTable(); // Keep TextTable only for now
-
+public class RandomVectorsTabularBenchmark extends AbstractVectorsBenchmark {
     @Param({"1000", "10000", "100000", "1000000"})
     int numBaseVectors;
 
@@ -54,33 +43,8 @@ public class RandomVectorsTabularBenchmark {
 
     @Setup
     public void setup() throws IOException {
-        originalDimension = 128;
-
-        baseVectors = new ArrayList<>(numBaseVectors);
-        queryVectors = new ArrayList<>(numQueryVectors);
-
-        for (int i = 0; i < numBaseVectors; i++) {
-            VectorFloat<?> vector = createRandomVector(originalDimension);
-            baseVectors.add(vector);
-        }
-
-        for (int i = 0; i < numQueryVectors; i++) {
-            VectorFloat<?> vector = createRandomVector(originalDimension);
-            queryVectors.add(vector);
-        }
-
-        // wrap the raw vectors in a RandomAccessVectorValues
-        ravv = new ListRandomAccessVectorValues(baseVectors, originalDimension);
-        // score provider using the raw, in-memory vectors
-        BuildScoreProvider bsp = BuildScoreProvider.randomAccessScoreProvider(ravv, VectorSimilarityFunction.EUCLIDEAN);
-
-        graphIndexBuilder = new GraphIndexBuilder(bsp,
-                ravv.dimension(),
-                16, // graph degree
-                100, // construction search depth
-                1.2f, // allow degree overflow during construction by this factor
-                1.2f); // relax neighbor diversity requirement by this factor
-        graphIndex = graphIndexBuilder.build(ravv);
+        tableRepresentation = new TextTable();
+        commonSetupRandom(numBaseVectors, numQueryVectors);
 
         transactionCount.set(0);
         totalLatency.set(0);
@@ -97,31 +61,9 @@ public class RandomVectorsTabularBenchmark {
             double meanLatency = (totalTransactions > 0) ? (double) totalLatency.get() / totalTransactions : 0.0;
             double p999Latency = calculateP999Latency();
             double meanVisited = (totalTransactions > 0) ? (double) visitedSamples.stream().mapToInt(Integer::intValue).sum() / totalTransactions : 0.0;
-            double recall = (totalTransactions > 0) ? (double) recallSamples.stream().mapToLong(Long::longValue).sum() / totalTransactions : 0.0;
+            double recall = (totalTransactions > 0) ? recallSamples.stream().mapToDouble(Double::doubleValue).sum() / totalTransactions : 0.0;
             tableRepresentation.addEntry(elapsed, count, meanLatency, p999Latency, meanVisited, recall);
         }, 1, 1, TimeUnit.SECONDS);
-    }
-
-    private VectorFloat<?> createRandomVector(int dimension) {
-        VectorFloat<?> vector = VECTOR_TYPE_SUPPORT.createFloatVector(dimension);
-        for (int i = 0; i < dimension; i++) {
-            vector.set(i, (float) Math.random());
-        }
-        return vector;
-    }
-
-    private double calculateP999Latency() {
-        if (latencySamples.isEmpty()) return 0.0;
-
-        List<Long> sortedLatencies = new ArrayList<>(latencySamples);
-        Collections.sort(sortedLatencies);
-
-        int p_count = sortedLatencies.size() / 1000;
-        int start_index = Math.max(0, sortedLatencies.size() - p_count);
-        int end_index = sortedLatencies.size() - 1;
-        List<Long> subList = sortedLatencies.subList(start_index, end_index);
-        long total = subList.stream().mapToLong(Long::longValue).sum();
-        return (double) total / p_count;
     }
 
     @TearDown
