+++ benchmarks-jmh/src/main/java/io/github/jbellis/jvector/bench/StaticSetVectorsBenchmark.java
@@ -18,10 +18,8 @@ package io.github.jbellis.jvector.bench;
 import io.github.jbellis.jvector.example.SiftSmall;
 import io.github.jbellis.jvector.example.util.SiftLoader;
 import io.github.jbellis.jvector.graph.*;
-import io.github.jbellis.jvector.graph.similarity.BuildScoreProvider;
 import io.github.jbellis.jvector.util.Bits;
 import io.github.jbellis.jvector.vector.VectorSimilarityFunction;
-import io.github.jbellis.jvector.vector.types.VectorFloat;
 import org.openjdk.jmh.annotations.*;
 import org.openjdk.jmh.infra.Blackhole;
 import org.slf4j.Logger;
@@ -39,15 +37,9 @@ import java.util.concurrent.TimeUnit;
 @Warmup(iterations = 2)
 @Measurement(iterations = 5)
 @Threads(1)
-public class StaticSetVectorsBenchmark {
+public class StaticSetVectorsBenchmark extends AbstractVectorsBenchmark {
     private static final Logger log = LoggerFactory.getLogger(StaticSetVectorsBenchmark.class);
-    private RandomAccessVectorValues ravv;
-    private ArrayList<VectorFloat<?>> baseVectors;
-    private ArrayList<VectorFloat<?>> queryVectors;
     private ArrayList<Set<Integer>> groundTruth;
-    private GraphIndexBuilder graphIndexBuilder;
-    private GraphIndex graphIndex;
-    int originalDimension;
 
     @Setup
     public void setup() throws IOException {
@@ -55,22 +47,7 @@ public class StaticSetVectorsBenchmark {
         baseVectors = SiftLoader.readFvecs(String.format("%s/siftsmall_base.fvecs", siftPath));
         queryVectors = SiftLoader.readFvecs(String.format("%s/siftsmall_query.fvecs", siftPath));
         groundTruth = SiftLoader.readIvecs(String.format("%s/siftsmall_groundtruth.ivecs", siftPath));
-        log.info("base vectors size: {}, query vectors size: {}, loaded, dimensions {}",
-                baseVectors.size(), queryVectors.size(), baseVectors.get(0).length());
-        originalDimension = baseVectors.get(0).length();
-        // wrap the raw vectors in a RandomAccessVectorValues
-        ravv = new ListRandomAccessVectorValues(baseVectors, originalDimension);
-
-        // score provider using the raw, in-memory vectors
-        BuildScoreProvider bsp = BuildScoreProvider.randomAccessScoreProvider(ravv, VectorSimilarityFunction.EUCLIDEAN);
-
-        graphIndexBuilder = new GraphIndexBuilder(bsp,
-                ravv.dimension(),
-                16, // graph degree
-                100, // construction search depth
-                1.2f, // allow degree overflow during construction by this factor
-                1.2f); // relax neighbor diversity requirement by this factor
-        graphIndex = graphIndexBuilder.build(ravv);
+        commonSetupStatic();
     }
 
     @TearDown
